sorunlarcozuldu
sorunlar çözüldü

+++ Kutuphane_Otomasyon_Taslak_winform/EmanetVer.cs
@@ -209,19 +209,31 @@ namespace Kutuphane_Otomasyon_Taslak_winform
         private void btnsil_Click(object sender, EventArgs e)
         {
             connection.Open();
-            DialogResult dialog;
-            dialog = MessageBox.Show("Bu Sepeti Boşaltmak İstiyor Musunuz?", "SİL!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-            string sil = "delete from sepet where kitapId='"+dataGridView1.CurrentRow.Cells["kitapId"].Value.ToString()+"'";
-            MySqlCommand command = new MySqlCommand(sil, connection);
-          
-            command.ExecuteNonQuery();
-            connection.Close();
+         
 
-            MessageBox.Show("Silme işlemi gerçekleşti.");
-          //  daset.Tables["sepet"].Clear();
-            sepetlistele();
-            lblkitapsayisi.Text = "";
-            kitapsayisi();
+
+            if (dataGridView1.CurrentRow==null)
+            {
+                MessageBox.Show("lütfen sepetteki silincek veriyi seçiniz");
+                connection.Close();
+            }
+            else
+            {
+                DialogResult dialog;
+                dialog = MessageBox.Show("Bu Sepeti Boşaltmak İstiyor Musunuz?", "SİL!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                string sil = "delete from sepet where kitapId='" + dataGridView1.CurrentRow.Cells["kitapId"].Value.ToString() + "'";
+                MySqlCommand command = new MySqlCommand(sil, connection);
+
+                command.ExecuteNonQuery();
+                connection.Close();
+
+                MessageBox.Show("Silme işlemi gerçekleşti.");
+                //  daset.Tables["sepet"].Clear();
+                sepetlistele();
+                lblkitapsayisi.Text = "";
+                kitapsayisi();
+            }
+
 
         }
